Update auth.py - Better Fix

+++ App/views/auth.py
@@ -30,15 +30,16 @@ def init():
 
     guess_digits_str = ''.join(guess_digits)
 
-    if guess_digits_str == sensei:
-        db.drop_all()
-        db.create_all()
-        create_user('bob', 'bobpass')
-        return render_template('201.html'), 201
+    if guess_digits_str:
+        if guess_digits_str == sensei:
+            db.drop_all()
+            db.create_all()
+            create_user('bob', 'bobpass')
+            return render_template('201.html'), 201
+        else:
+            return render_template('401.html'), 401
     else:
         return render_template('401.html'), 401
-        
-    return render_template('401.html'), 401
 
 
 @auth_views.route('/users', methods=['GET'])
